requirements.txt
faiss-cpu >= 1.6.3
datasets >= 1.0.1
torch >= 1.4.0
transformers